refactor: 컴포넌트 폴더 파일명 스네이크 케이스로 수정
