+++ src/screens/team/team_join/list/components/member_count_slider.tsx
