+++ src/screens/team/team_join/list/components/team_card.tsx
