+++ src/screens/team/team_join/list/components/team_list_header.tsx
