+++ src/screens/team/team_join/university/components/university_item.tsx
